+++ alitheia/core/src/main/java/eu/sqooss/impl/service/updater/UpdaterJob.java
@@ -51,7 +51,7 @@ public class UpdaterJob extends Job {
     }
     
     @Override
-    public int priority() {
+    public long priority() {
         return 0;
     }
